+++ src/processingunit/cpu/cpu2d.cpp
@@ -22,7 +22,7 @@ void CPU_2d::computeBorder(func_ptr_t* mUserFuncs, unsigned short int* mCompFunc
 		for (int x = 0; x < xCount; ++x) {
 			int xShift = x;
 			for (int y = 0; y < haloSize; ++y) {
-				int yShift = xCount * y;
+				unsigned long long yShift = xCount * y;
 				//cout << "Calc y_" << y << " x_" << x << endl;
 				//printf("Calc y = %d, x = %d\n", y, x);
 				mUserFuncs[mCompFuncNumber[yShift + xShift]](result, source, time, x, y, 0, parametrs, externalBorder);
@@ -33,7 +33,7 @@ void CPU_2d::computeBorder(func_ptr_t* mUserFuncs, unsigned short int* mCompFunc
 		for (int x = 0; x < xCount; ++x) {
 			int xShift = x;
 			for (int y = yCount - haloSize; y < yCount; ++y) {
-				int yShift = xCount * y;
+				unsigned long long yShift = xCount * y;
 				//cout << "Calc y_" << y << " x_" << x << endl;
 				//printf("Calc y = %d, x = %d\n", y, x);
 				mUserFuncs[mCompFuncNumber[yShift + xShift]](result, source, time, x, y, 0, parametrs, externalBorder);
@@ -42,7 +42,7 @@ void CPU_2d::computeBorder(func_ptr_t* mUserFuncs, unsigned short int* mCompFunc
 
 # pragma omp for
 		for (int y = haloSize; y < yCount - haloSize; ++y) {
-			int yShift = xCount * y;
+			unsigned long long yShift = xCount * y;
 			for (int x = 0; x < haloSize; ++x) {
 				int xShift = x;
 				//cout << "Calc y_" << y << " x_" << x << endl;
@@ -53,7 +53,7 @@ void CPU_2d::computeBorder(func_ptr_t* mUserFuncs, unsigned short int* mCompFunc
 
 # pragma omp for
 		for (int y = haloSize; y < yCount - haloSize; ++y) {
-			int yShift = xCount * y;
+			unsigned long long yShift = xCount * y;
 			for (int x = xCount - haloSize; x < xCount; ++x) {
 				int xShift = x;
 				//cout << "Calc y_" << y << " x_" << x << endl;
@@ -70,7 +70,7 @@ void CPU_2d::computeCenter(func_ptr_t* mUserFuncs, unsigned short int* mCompFunc
 	{
 # pragma omp for
 		for (int y = haloSize; y < yCount - haloSize; ++y) {
-			int yShift = xCount * y;
+			unsigned long long yShift = xCount * y;
 			for (int x = haloSize; x < xCount - haloSize; ++x) {
 				int xShift = x;
 				//cout << "Calc y_" << y << " x_" << x << endl;
